+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkStreamingDeploymentManagerProvider.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.engine.management
 
 import cats.data.ValidatedNel
 import com.typesafe.config.Config
+import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine._
 import pl.touk.nussknacker.engine.api.StreamMetaData
 import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
@@ -15,11 +16,12 @@ import pl.touk.nussknacker.engine.management.rest.FlinkClient
 import scala.concurrent.duration.FiniteDuration
 import scala.util.Try
 
-class FlinkStreamingDeploymentManagerProvider extends DeploymentManagerProvider {
+class FlinkStreamingDeploymentManagerProvider extends DeploymentManagerProvider with LazyLogging {
 
   import net.ceedubs.ficus.Ficus._
   import net.ceedubs.ficus.readers.ArbitraryTypeReader._
   import pl.touk.nussknacker.engine.util.config.ConfigEnrichments._
+  import ScenarioTestingConfig._
 
   override def createDeploymentManager(
       modelData: BaseModelData,
@@ -27,6 +29,7 @@ class FlinkStreamingDeploymentManagerProvider extends DeploymentManagerProvider 
       deploymentConfig: Config,
       scenarioStateCacheTTL: Option[FiniteDuration]
   ): ValidatedNel[String, DeploymentManager] = {
+    logger.info("Creating FlinkStreamingDeploymentManager")
     import dependencies._
     val flinkConfig = deploymentConfig.rootAs[FlinkConfig]
     FlinkClient.create(flinkConfig, scenarioStateCacheTTL).map { client =>